switch to clang, code fixes (to shut compiler up)

+++ src/include/file.h
@@ -5,7 +5,7 @@
 
 char *get_file_path(const char *type);
 bool file_exists(const char *filename);
-char **read_file(const char *filename, int *lines_read);
+char **read_file(const char *filename, unsigned long *lines_read);
 int write_file(const char *filename, const char *content, const char *mode);
 
 #endif /* FILE_H */
